# Add support for multi-threaded fine rasterization