--- a/sparse_strips/vello_bench/Cargo.toml
+++ b/sparse_strips/vello_bench/Cargo.toml
@@ -18,6 +18,12 @@ rand = { workspace = true }
 smallvec = { workspace = true }
 usvg = { workspace = true }
 
+[features]
+# We need this because the `pack` benchmark cannot be run with multithreading enabled, but
+# `cargo nextest` in CI will attempt to do so. So we disable the benchmark when that feature
+# is enabled.
+multithreading = ["vello_cpu/multithreading"]
+
 [[bench]]
 name = "main"
 harness = false