--- a/sparse_strips/vello_bench/src/fine/mod.rs
+++ b/sparse_strips/vello_bench/src/fine/mod.rs
@@ -7,35 +7,22 @@ mod gradient;
 mod image;
 mod rounded_blurred_rect;
 mod strip;
-
-use criterion::Bencher;
-use vello_cpu::fine::{Fine, FineType, SCRATCH_BUF_SIZE};
-use vello_dev_macros::vello_bench;
+// CI will attempt to build this crate with all features enabled, but the problem
+// is that the `update_regions` function has a slightly different signature with multithreading
+// enabled, which makes it incompatible with the `Bencher` closure. Because of this, we add
+// this feature to `vello_bench` as well and disable the benchmark in case it's enabled.
+#[cfg(not(feature = "multithreading"))]
+mod pack;
 
 pub use blend::*;
 pub use fill::*;
 pub use gradient::*;
 pub use image::*;
+#[cfg(not(feature = "multithreading"))]
+pub use pack::*;
 pub use rounded_blurred_rect::*;
 pub use strip::*;
-use vello_common::coarse::WideTile;
 use vello_common::peniko::{BlendMode, Compose, Mix};
-use vello_common::tile::Tile;
-use vello_cpu::region::Regions;
-
-#[cfg(not(feature = "multithreading"))]
-#[vello_bench]
-pub fn pack<F: FineType>(b: &mut Bencher<'_>, fine: &mut Fine<F>) {
-    let mut buf = vec![0; SCRATCH_BUF_SIZE];
-    let mut regions = Regions::new(WideTile::WIDTH, Tile::HEIGHT, &mut buf);
-
-    b.iter(|| {
-        regions.update_regions(|r| {
-            fine.pack(r);
-            std::hint::black_box(&r);
-        });
-    });
-}
 
 pub(crate) fn default_blend() -> BlendMode {
     BlendMode::new(Mix::Normal, Compose::SrcOver)