--- a/sparse_strips/vello_cpu/src/lib.rs
+++ b/sparse_strips/vello_cpu/src/lib.rs
@@ -113,6 +113,8 @@ mod render;
 #[doc(hidden)]
 /// This is an internal module, do not access directly.
 pub mod fine;
+#[doc(hidden)]
+pub mod region;
 mod util;
 
 pub use render::RenderContext;