--- a/sparse_strips/vello_dev_macros/src/bench.rs
+++ b/sparse_strips/vello_dev_macros/src/bench.rs
@@ -36,12 +36,12 @@ pub(crate) fn vello_bench_inner(_: TokenStream, item: TokenStream) -> TokenStrea
             }
 
             c.bench_function(&get_bench_name(&#input_fn_name_str, "u8_scalar"), |b| {
-                let mut fine = Fine::<u8>::new(WideTile::WIDTH, Tile::HEIGHT);
+                let mut fine = Fine::<u8>::new();
                 #inner_fn_name(b, &mut fine);
             });
 
             c.bench_function(&get_bench_name(&#input_fn_name_str, "f32_scalar"), |b| {
-                let mut fine = Fine::<f32>::new(WideTile::WIDTH, Tile::HEIGHT);
+                let mut fine = Fine::<f32>::new();
                 #inner_fn_name(b, &mut fine);
             });
         }